rename test files to include '.test.'
