+++ test/integration/lists.test.ts
@@ -1,7 +1,7 @@
 import test from 'ava';
 import { rest } from 'msw';
 import { DiscogsClient } from '@lib/client.js';
-import { setupMockAPI } from './_setup.js';
+import { setupMockAPI } from './_setup.test.js';
 
 const server = setupMockAPI();
