+++ test/unit/error.test.ts
